Flask/requirements.txt
numpy==1.18.5
Flask==1.1.1
Werkzeug==0.16.1
tensorflow==2.1.0
opencv_python
tensorflow-gpu==2.10.0
pillow
scikit-learn